% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\title{Jackknife}
\usage{
jackknife(.x, .f, ...)
}
\arguments{
\item{.x}{A vector of values.}

\item{.f}{A function that calculates a summary statistic of \code{.x}.}

\item{...}{Additional arguments passed to \code{.f}.}
}
\description{
Calculate the jackknife standard error of a function \code{.f} using jackknife resampling.
}
\examples{
x <- rnorm(100)
jackknife(.x = x, .f = mean)
}
