% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.mcerror-package.R
\docType{package}
\name{mice.mcerror_package}
\alias{mice.mcerror_package}
\title{Calculate Monte Carlo Error Estimates for Statistics Computed Using Multiply Imputed Data}
\description{
Calculate Monte Carlo error estimates for statistics computed by pooling multiply imputed data. This package is an add-on to the {mice} package, which is thus required.
}
\author{
Alessandro Gasparini (alessandro@ellessenne.xyz)
}
